created EDA-int branch
This branch aims to perform Driven Molecular Dynamics simulation within the Electric Dipola Approximation, i.e. by the uese of an external time-dependent homogeneous electric field. The pursued way will be the implementation of a suitable integrator, differently from eda branch where a new ForceField was intented to be implemented.